refactor(modules): Split up src/utils/dom.js
