+++ src/utils/dom/animationEndEvent.js
@@ -1,0 +1,22 @@
+import { isNodeEnv } from '../isNodeEnv'
+
+export const animationEndEvent = (() => {
+  // Prevent run in Node env
+  if (isNodeEnv()) {
+    return false
+  }
+
+  const testEl = document.createElement('div')
+  const transEndEventNames = {
+    'WebkitAnimation': 'webkitAnimationEnd',
+    'OAnimation': 'oAnimationEnd oanimationend',
+    'animation': 'animationend'
+  }
+  for (const i in transEndEventNames) {
+    if (transEndEventNames.hasOwnProperty(i) && typeof testEl.style[i] !== 'undefined') {
+      return transEndEventNames[i]
+    }
+  }
+
+  return false
+})()
