+++ src/utils/isNodeEnv.js
@@ -1,0 +1,2 @@
+// Detect Node env
+export const isNodeEnv = () => typeof window === 'undefined' || typeof document === 'undefined'
